feat: implement auth check in expenses features

+++ routes/expenses.js
@@ -12,10 +12,11 @@ router.get('/', (req, res, next) => {
   const page = parseInt(req.query.page) || 1
   const offset = (page - 1) * limit
   const sort = req.query.sort || ''
+  const userId = req.user.id
 
-  let whereCondition = {};
+  let whereCondition = { userId };
   if (sort) {
-    whereCondition = { categoryId: parseInt(sort) }; // Assuming sort value is category ID
+    whereCondition = { ...whereCondition, categoryId: parseInt(sort) }; // Assuming sort value is category ID
   }
 
   return Promise.all([
@@ -64,6 +65,8 @@ router.get('/', (req, res, next) => {
 
 // Get New Page
 router.get('/new', (req, res, next) => {
+  const user = req.user
+
   return Category.findAll({
     attributes: [`id`, `name`, `icon`],
     raw: true
@@ -71,7 +74,7 @@ router.get('/new', (req, res, next) => {
     .then((categories) => {
       const { sort, page } = req.query
       const initialCategory = categories.length > 0 ? categories[0].id : null;
-      return res.render('new', { categories, initialCategory, sort, page })
+      return res.render('new', { categories, initialCategory, sort, page, user })
     })
     .catch((error) => {
       error.errorMessage = 'Failed to get this page! :('
@@ -83,9 +86,11 @@ router.get('/new', (req, res, next) => {
 router.get('/:id/edit', (req, res, next) => {
   const id = req.params.id
   const { sort, page } = req.query
+  const userId = req.user.id
+
   return Promise.all([
     Expense.findByPk(id, {
-      attributes: [`id`, `name`, `date`, `amount`, `categoryId`],
+      attributes: [`id`, `name`, `date`, `amount`, `categoryId`, `userId`],
       raw: true
     }),
     Category.findAll({
@@ -94,10 +99,17 @@ router.get('/:id/edit', (req, res, next) => {
     })
   ])
     .then(([expense, categories]) => {
+
       if (!expense) {
         req.flash('error', 'Cannot find any data! :(')
         return res.redirect('/expenses')
       }
+
+      if (expense.userId !== userId) {
+        req.flash('error', 'You are not authorized to edit this expense! :(')
+        return res.redirect('/expenses')
+      }
+
       expense.date = new Date(expense.date).toISOString().split('T')[0]
       const initialCategory = expense.categoryId
       return res.render('edit', { expense, categories, initialCategory, sort, page })
@@ -112,7 +124,9 @@ router.get('/:id/edit', (req, res, next) => {
 router.post('/', (req, res, next) => {
   const { name, date, amount, category } = req.body
   const { sort, page } = req.query
-  return Expense.create({ name, date, amount, categoryId: category })
+  const userId = req.user.id
+
+  return Expense.create({ name, date, amount, categoryId: category, userId })
     .then(() => {
       req.flash('success', 'Created successfully!')
       return res.redirect(`/expenses?sort=${sort}&page=${page}`)
@@ -128,10 +142,28 @@ router.put('/:id', (req, res, next) => {
   const id = req.params.id
   const { sort, page } = req.query
   const { name, date, amount, category } = req.body
-  return Expense.update({ name, date, amount, categoryId: category }, { where: { id } })
-    .then(() => {
-      req.flash('success', 'Updated successfully!')
-      return res.redirect(`/expenses?sort=${sort}&page=${page}`)
+  const userId = req.user.id
+
+  return Expense.findByPk(id, {
+    attributes: [`id`, `name`, `date`, `amount`, `categoryId`, `userId`],
+    raw: true
+  })
+    .then((expense) => {
+      if (!expense) {
+        req.flash('error', 'Cannot find any data! :(')
+        return res.redirect('/expenses')
+      }
+
+      if (expense.userId !== req.user.id) {
+        req.flash('error', 'You are not authorized to edit this expense! :(')
+        return res.redirect('/expenses')
+      }
+
+      return Expense.update({ name, date, amount, categoryId: category }, { where: { id } })
+        .then(() => {
+          req.flash('success', 'Updated successfully!')
+          return res.redirect(`/expenses?sort=${sort}&page=${page}`)
+        })
     })
     .catch((error) => {
       error.errorMessage = 'Failed to update this expense! :('
@@ -143,10 +175,29 @@ router.put('/:id', (req, res, next) => {
 router.delete('/:id', (req, res, next) => {
   const id = req.params.id
   const { sort, page } = req.query
-  return Expense.destroy({ where: { id } })
-    .then(() => {
-      req.flash('success', 'Deleted successfully!')
-      return res.redirect(`/expenses?sort=${sort}&page=${page}`)
+  const userId = req.user.id
+
+  return Expense.findByPk(id, {
+    attributes: [`id`, `name`, `date`, `amount`, `categoryId`, `userId`],
+    raw: true
+  })
+    .then((expense) => {
+
+      if (!expense) {
+        req.flash('error', 'Cannot find any data! :(')
+        return res.redirect('/expenses')
+      }
+
+      if (expense.userId !== req.user.id) {
+        req.flash('error', 'You are not authorized to delete this expense! :(')
+        return res.redirect('/expenses')
+      }
+
+      return Expense.destroy({ where: { id } })
+        .then(() => {
+          req.flash('success', 'Deleted successfully!')
+          return res.redirect(`/expenses?sort=${sort}&page=${page}`)
+        })
     })
     .catch((error) => {
       error.errorMessage = 'Failed to delete this expense! :('
